+++ src/components/index.js
@@ -1,3 +1,5 @@
 export * from './marginals';
 
-export { default as AddJournal} from './AddJournal/AddJournal'
+export { default as AddJournal} from './AddJournal/AddJournal';
+export { default as Login } from './Authentication/Login/Login';
+export { default as SignUp } from './Authentication/Signup/SignUp'
